create src folder project
