+++ .vscode/settings.json
@@ -5,5 +5,13 @@
             "contentType": "application/json",
             "X-Auth-Token": "169ddeb1-502a-42cf-a222-9dbb8ec2cbf6",
         }
-    }
+    },
+    "python.analysis.extraPaths": [
+        "./src"
+    ],
+    "python.testing.pytestArgs": [
+        "tests"
+    ],
+    "python.testing.unittestEnabled": false,
+    "python.testing.pytestEnabled": true
 }
