--- requirements.dev.in
@@ -1,3 +1,0 @@
--r requirements.in
-pytest
-pip-tools
