+++ tests/test_main.py
@@ -1,6 +1,5 @@
 import sys
 import os
-import pytest
 from fastapi.testclient import TestClient
 from unittest.mock import patch
